==> 01_Create Postgres Database.ipynb <==
import pandas as pd

#get pandas and sql to work together
import psycopg2 as pg
import pandas.io.sql as pd_sql

from psycopg2 import connect
from psycopg2.extensions import ISOLATION_LEVEL_AUTOCOMMIT

params = {
    'host': 'localhost',  # We are connecting to our _local_ version of psql
    'user': 'agar',
    'dbname': 'votes',    # DB that we are connecting to
    'port': 5432          # port we opened on AWS
}

# We will talk about this magic Python trick!
connection = pg.connect(**params)

ls

# Connect and create database, disconnect, and reconnect to the right database
connection = connect(**params, dbname='agar')
connection.set_isolation_level(ISOLATION_LEVEL_AUTOCOMMIT)
connection.cursor().execute('CREATE DATABASE votes;')
connection.close()

df.head()

# Postgres info to connect???

connection_args = {
    'host': 'localhost',  # We are connecting to our _local_ version of psql
    'dbname': 'names',    # DB that we are connecting to
    'port': 5432          # port we opened on AWS
}

# We will talk about this magic Python trick!
connection = pg.connect(**connection_args)

df.iloc[:0]

votes_df = pd.read_csv("/Users/agar/_METIS/exercises/Project_3/data_source/1976-2018-house.csv", engine='python')


votes_df.head()

# This is a different way of creating tables
# It is convinient if there are many columns
from sqlalchemy import create_engine
connection_string = f'postgres://agar:{params["host"]}@{params["host"]}:{params["port"]}/votes'
engine = create_engine(connection_string, pool_pre_ping=True)


votes_df.iloc[:0].to_sql("H_of_Rep", engine, index=False)

votes_df.iloc[:].to_sql("H_of_Rep", engine, index=False, if_exists = 'append', chunksize = 1000)

connection = connect(**params, dbname='votes')
cursor = connection.cursor()
cursor.execute("SELECT * FROM con_dist_votes LIMIT 1;")
cursor.fetchall()

cursor.execute("SELECT * FROM con_dist_votes LIMIT 1;")
cursor.fetchall()

demographic_df = pd.read_csv("/Users/agar/_METIS/exercises/Project_3/data_source/115th Congress Members Guide with Elections and Demographic Data by District - House.csv")

demographic_dfs = demographic_df.iloc[1:]

demographic_dfs.head()

demographic_dfs.iloc[:0]

demographic_dfs.columns = ['District', 'Code', 'Party', 'Incumbent Name', 'Unnamed: 4',
       'Pronunciation Guide', 'Unnamed: 6', 'First Elected', 'Birth Year',
       'Gender', 'Race/ Ethnicity', 'Religion', 'LGBTQ', 'Pre-2016 Incumbent',
       '2016 President', 'Unnamed: 15', '2012 President', 'Unnamed: 17',
       '2008 President', 'Unnamed: 19', '2016 US House', 'Unnamed: 21',
       '2014 US House', 'Unnamed: 23', '2012 Two-Party', 'Unnamed: 25',
       '2012-2016 ACS Citizen Adult Population', 'Unnamed: 27', 'Unnamed: 28',
       'Unnamed: 29', 'Unnamed: 30', 'Unnamed: 31',
       '2010 Census Adult Population by Race', 'Unnamed: 33', 'Unnamed: 34',
       'Unnamed: 35', 'Unnamed: 36', 'Unnamed: 37',
       '2016 American Community Survey Income and Education', 'Unnamed: 39',
       'Unnamed: 40', 'Unnamed: 41', 'Unnamed: 42', 'Unnamed: 43',
       'Non-College White Share', '2013-2018 House Specials', 'Unnamed: 46',
       'Unnamed: 47', 'Unnamed: 48', '2016 President.1', 'Unnamed: 50',
       'Unnamed: 51', '2012 President.1', 'Unnamed: 53', 'Unnamed: 54',
       '2008 President.1', 'Unnamed: 56', 'Unnamed: 57', '2016 House',
       'Unnamed: 59', 'Unnamed: 60', 'Unnamed: 61', '2014 House',
       'Unnamed: 63', 'Unnamed: 64', 'Unnamed: 65', '2012 House 2-Party',
       'Unnamed: 67', 'Special Elections', 'Unnamed: 69', 'Unnamed: 70',
       'Unnamed: 71', 'ACS Citizen Population_Includes Latinos of Any Race',
       'Unnamed: 73', 'Unnamed: 74', 'Unnamed: 75', 'Unnamed: 76',
       'Unnamed: 77', 'Unnamed: 78',
       'ACS Citizen Voting Age Population_Includes Latinos of Any Race',
       'Unnamed: 80', 'Unnamed: 81', 'Unnamed: 82', 'Unnamed: 83',
       'Unnamed: 84', 'Unnamed: 85',
       'Census Total Population_Includes Latinos of Any Race', 'Unnamed: 87',
       'Unnamed: 88', 'Unnamed: 89', 'Unnamed: 90', 'Unnamed: 91',
       'Unnamed: 92',
       'Census Voting Age Population_Includes Latinos of Any Race',
       'Unnamed: 94', 'Unnamed: 95', 'Unnamed: 96', 'Unnamed: 97',
       'Unnamed: 98', 'Unnamed: 99']

demographic_dfs.columns

demographic_dfs.iloc[:0].to_sql("demographic_115_con", engine, index=False, if_exists = 'append', chunksize = 1000)

demographic_dfs.to_sql("demographic_115_con", engine, index=False, if_exists = 'append', chunksize = 1000)

geography_df = pd.read_csv("/Users/agar/_METIS/exercises/Project_3/data_source/Congressional district geographic descriptions and largest places - 116th Congress - By place.csv")

geography_df.columns = ['District', 'Incumbent', 'Party', 'Geographic Description',
       'Largest place', 'p_of CD', 'Second-largest place', 'p_of CD.1',
       'Third-largest place', 'p_of CD.2', 'p_None']

geography_df.head()

geography_df.dtypes.columns

geography_df.iloc[:0].to_sql("geography_116+", engine, index=False, if_exists = 'append', chunksize = 1000)

geography_df.to_sql("geography_116+", engine, index=False, if_exists = 'append', chunksize = 1000)

geography_metro_df = pd.read_csv("/Users/agar/_METIS/exercises/Project_3/data_source/Congressional district geographic descriptions and largest places - 116th Congress - By metro area.csv")

geography_metro_df.columns = ['District', 'Largest metro area', 'p_of CD',
       'Second-largest metro area', 'p_of CD.1', 'Third-largest metro area',
       'p_of CD.2', 'p_no metro']

geography_metro_df.head()

geography_metro_df[:0].to_sql("geography_116_metro", engine, index=False, if_exists = 'append', chunksize = 1000)

geography_metro_df.to_sql("geography_116_metro", engine, index=False, if_exists = 'append', chunksize = 1000)

#grab financial data
df_2014 = pd.read_csv("/Users/agar/_METIS/exercises/Project_3/data_source/candidate_money/candidate_summary_2014.csv")
df_2016 = pd.read_csv("/Users/agar/_METIS/exercises/Project_3/data_source/candidate_money/candidate_summary_2016.csv")
df_2018 = pd.read_csv("/Users/agar/_METIS/exercises/Project_3/data_source/candidate_money/candidate_summary_2018.csv")

df_2014['year'] = 2014
df_2016['year'] = 2016
df_2018['year'] = 2018

df_spending = pd.concat([df_2014, df_2016, df_2018], ignore_index=True)

#send to postgres
df_spending.iloc[:0].to_sql("financials", engine, index=False, if_exists = 'append', chunksize = 1000)
df_spending.to_sql("financials", engine, index=False, if_exists = 'append', chunksize = 1000)
