==> 02_SQL_query_Data_merge.ipynb <==
import pandas as pd

#get pandas and sql to work together
import psycopg2 as pg
import pandas.io.sql as pd_sql

from psycopg2 import connect
from psycopg2.extensions import ISOLATION_LEVEL_AUTOCOMMIT
import matplotlib.pyplot as plt

params = {
    'host': 'localhost',  # We are connecting to our _local_ version of psql
    'user': 'agar',
    'dbname': 'votes',    # DB that we are connecting to
    'port': 5432          # port we opened on AWS
}

# We will talk about this magic Python trick!
connection = pg.connect(**params)

sql_query_demographics = """
select 
"Code" as district_id, 
cast("2010 Census Adult Population by Race" as float) as white_adult_percent, 
cast("Unnamed: 33" as float) as black_adult_percent, 
cast("Unnamed: 34" as float) as latino_adult_percent,
cast("Unnamed: 35" as float) as asian_adult_percent,
cast("Unnamed: 36" as float) as native_adult_percent,
cast("Unnamed: 37" as float) as other_adult_percent,
cast("2016 American Community Survey Income and Education" as float) as bachelors,
cast("Unnamed: 40" as float) as white_bachelors,
cast("Non-College White Share" as float) as white_no_college,
cast (REPLACE(replace("Unnamed: 42", ',', ''), '$', '') as integer) as income_median,
cast(replace("Census Voting Age Population_Includes Latinos of Any Race", ',', '') as integer) as total_vote,
cast(replace("Unnamed: 94", ',', '') as integer) as white_vote,
cast(replace("Unnamed: 95", ',', '') as integer) as black_vote,
cast(replace("Unnamed: 96", ',', '') as integer) as latino_vote,
cast(replace("Unnamed: 97", ',', '') as integer) as asian_vote,
cast(replace("Unnamed: 98", ',', '') as integer) as native_vote,
cast(replace("Unnamed: 99", ',', '') as integer) as other_vote,
cast(replace("Census Total Population_Includes Latinos of Any Race", ',', '') as integer) as total_pop,
cast(replace("Unnamed: 87", ',', '') as integer) as white_pop,
cast(replace("Unnamed: 88", ',', '') as integer) as black_pop,
cast(replace("Unnamed: 89", ',', '') as integer) as latino_pop,
cast(replace("Unnamed: 90", ',', '') as integer) as asian_pop,
cast(replace("Unnamed: 91", ',', '') as integer) as native_pop,
cast(replace("Unnamed: 92", ',', '') as integer) as other_pop

from "demographic_115_con" 
"""

sql_query_largest_geo = """
select 
"District" as district_id,
"p_of CD" as larg1_percent_pop,
"p_of CD.1" as larg2_percent_pop,
"p_of CD.2" as larg3_percent_pop
from "geography_116+" g 
"""

sql_query_metro = """
select 
"District" as district_id,
"p_of CD" as metro1_percent_pop,
"p_of CD.1" as metro2_percent_pop,
"p_of CD.2" as metro3_percent_pop,
"p_no metro" as metro_none_percent_pop
from "geography_116_metro"  
"""

sql_query_votes = """
select year,  
case 
	when district<10 
		then concat(state_po, '-0', district)
	when district>9	
		then concat(state_po, '-', district)
		end as district_id,
state_fips, party,
cast(replace(candidatevotes, ',', '') as integer) as candidate_votes, 
totalvotes

from "H_of_Rep" hor 

where year>2009 and year<2020
order by year, district_id, party
"""

df_demographics = pd.read_sql_query(sql_query_demographics, connection, index_col=None, coerce_float=True, params=None, parse_dates=None, chunksize=None)

df_demographics.head(440)

df_largest_geo = pd.read_sql_query(sql_query_largest_geo, connection, index_col=None, coerce_float=True, params=None, parse_dates=None, chunksize=None)

df_largest_geo.head()

df_metro = pd.read_sql_query(sql_query_metro, connection, index_col=None, coerce_float=True, params=None, parse_dates=None, chunksize=None)

df_metro.head()

df_metro["district_id"][0] = "AK-00"

district_merge_1 = pd.merge(df_metro, df_largest_geo, how='inner', on="district_id")

district_merge_1b = pd.merge(district_merge_1, df_demographics, how='inner', on="district_id")

district_merge_1b.head()

new = district_merge_1["district_id"].str.split("-", n = 1, expand = True) 

new.head()

new_num = new[1].replace("AL", '00')

new_num.head()

news = pd.concat([new[0], new_num], axis=1, sort=False)

news_district_fix = news[0] + "-" + news[1]

news_district_fix

district_merge_2 = pd.concat([news_district_fix, district_merge_1b], axis=1, sort=False)

district_merge_2 = district_merge_2.drop(columns = "district_id")

district_merge_2.columns

district_merge_2 = district_merge_2.rename(columns={0: "district_id"})

district_merge_2.head()

df_votes = pd.read_sql_query(sql_query_votes, connection, index_col=None, coerce_float=True, params=None, parse_dates=None, chunksize=None)

df_votes.head()

df_votes.tail()

sql_query_winners_ten = """
select year, party, district_id
from "Winners+" w 
where year>2009 and year<2020
order by year, district_id, party
"""

df_winners_ten = pd.read_sql_query(sql_query_winners_ten, connection, index_col=None, coerce_float=True, params=None, parse_dates=None, chunksize=None)

df_winners_ten

winner_votes = pd.merge(df_winners_ten, df_votes, how='left', on=["year", "district_id", "party"])

sql_query_flips_agg = """
select * from
(
select *, 
sum(party_change_simple) over (partition by district_id order by year asc) as sum_flips_total
from "Flips" 
) as total_flips
where year>2009 and year<2019 
"""

df_flips_agg = pd.read_sql_query(sql_query_flips_agg, connection, index_col=None, coerce_float=True, params=None, parse_dates=None, chunksize=None)

df_flips_agg.head()

winner_votes_agg = pd.merge(df_flips_agg, winner_votes, how='left', on=["year", "district_id"])

winner_votes_agg.head()

complete_df = pd.merge(winner_votes_agg, district_merge_2, how='left', on=["district_id"])

complete_df.head()

complete_df.info()

complete_df = complete_df.fillna(0)

complete_df.to_csv('/Users/agar/_METIS/exercises/Project_3/data_source/complete_df') 

from sqlalchemy import create_engine
connection_string = f'postgres://agar:{params["host"]}@{params["host"]}:{params["port"]}/votes'
engine = create_engine(connection_string, pool_pre_ping=True)

complete_df.iloc[:0].to_sql("Complete", engine, index=False)
complete_df.iloc[:].to_sql("Complete", engine, index=False, if_exists = 'append', chunksize = 1000)

 sql_query_finance = """
select * from "financials"
"""

df_financials = pd.read_sql_query(sql_query_finance, connection, index_col=None, coerce_float=True, params=None, parse_dates=None, chunksize=None)

#get only House of Representatives
mask = df_financials['Cand_Office'] == "H"
house_df = df_financials[mask]

#select financial columns 
df_in = house_df[['Cand_State', 'Cand_Office_Dist','year', 'Total_Receipt', 'Total_Disbursement', 'Individual_Itemized_Contribution',
       'Individual_Unitemized_Contribution', 'Individual_Contribution',
       'Other_Committee_Contribution', 'Party_Committee_Contribution',
       'Cand_Contribution', 'Total_Contribution', 'Operating_Expenditure']]

#sum expenditures by district and year
df_financial_sum = df_in.groupby(['Cand_State', 'Cand_Office_Dist','year']).agg('sum')
df_financial_sum_r = df_financial_sum.reset_index()

#set columns for joins
df_financial_sum_r['Cand_Office_Dist'] = df_financial_sum_r.Cand_Office_Dist.map("{:02}".format)
df_financial_sum_r['ID_DIST'] =  df_financial_sum_r['year'].astype(str)+ "-" + df_financial_sum_r['Cand_State'] + "-" + df_financial_sum_r['Cand_Office_Dist'].astype(str) 
df_financial_sum_r['district_id'] =  df_financial_sum_r['Cand_State']+ "-" + df_financial_sum_r['Cand_Office_Dist'].astype(str) 
df_r.to_csv("financials_df.csv")

df_r.to_csv("financials_df.csv")
